requirements-dev.txt
psycopg2==2.6.1
docker-py==1.7.0
